requirements.txt
# Core dependencies
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langgraph>=0.2.0,<0.3.0
pydantic>=2.0.0,<3.0.0
fastapi>=0.100.0,<0.120.0
uvicorn>=0.23.0,<0.30.0

# --- Let pip resolve these Google packages for compatibility ---
langchain-google-genai
google-generativeai

# Database and storage
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.5,<3.0.0
neo4j>=5.8.0,<6.0.0
elasticsearch>=8.9.0,<9.0.0
faiss-cpu>=1.7.4,<1.9.0
pinecone-client>=2.2.1,<5.0.0

# Vector embeddings
sentence-transformers>=2.2.2,<3.0.0
transformers>=4.30.0,<5.0.0
torch>=2.0.0,<2.5.0
tokenizers>=0.15.0,<1.0.0

# Data processing
arxiv>=1.4.7,<3.0.0
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.1.0
spacy>=3.6.0,<4.0.0
nltk>=3.8.1,<4.0.0

# PDF processing
PyMuPDF>=1.22.0,<2.0.0
pdfminer.six>=20221105,<20250000

# Visualization and UI
streamlit>=1.25.0,<2.0.0
plotly>=5.15.0,<6.0.0
gradio>=4.0.0,<5.0.0

# Essential supporting packages
langsmith>=0.4.0,<1.0.0
typing-extensions>=4.7.0,<5.0.0
httpx>=0.25.0,<1.0.0
orjson>=3.9.0,<4.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0
tqdm>=4.65.0,<5.0.0
pytest>=7.4.0,<9.0.0
black>=23.7.0,<25.0.0
isort>=5.12.0,<6.0.0
filetype>=1.2.0,<2.0.0